+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* [PR-228](https://github.com/itk-dev/economics/pull/228)
+  Use jobs to reduce synchronization load.
+
 ## [2.9.4] - 2025-07-10
 
 * [PR-232](https://github.com/itk-dev/economics/pull/232)
@@ -495,9 +498,7 @@ complete process.
 * Updated to authorization code flow.
 * Changed worklog save button styling to be sticky.
 
-[Unreleased]: https://github.com/itk-dev/economics/compare/2.9.4...HEAD
-[2.9.4]: https://github.com/itk-dev/economics/compare/2.9.3...2.9.4
-[2.9.3]: https://github.com/itk-dev/economics/compare/2.9.2...2.9.3
+[Unreleased]: https://github.com/itk-dev/economics/compare/2.9.2...HEAD
 [2.9.2]: https://github.com/itk-dev/economics/compare/2.9.0...2.9.2
 [2.9.0]: https://github.com/itk-dev/economics/compare/2.8.6...2.9.0
 [2.8.6]: https://github.com/itk-dev/economics/compare/2.8.5...2.8.6